Extract `getCertificateData`
